Refactor Canton integration-test-lib
