+++ daml-lf/integration-test-lib/BUILD.bazel
@@ -18,11 +18,7 @@ da_scala_library(
         "@maven//:org_scalaz_scalaz_core",
     ],
     scalacopts = lf_scalacopts_stricter,
-    visibility = [
-        "//daml-script:__subpackages__",
-        "//navigator:__subpackages__",
-        "//triggers:__subpackages__",
-    ],
+    visibility = ["//visibility:public"],
     deps = [
         "//bazel_tools/runfiles:scala_runfiles",
         "//daml-lf/archive:daml_lf_archive_reader",
