+++ daml-script/export/integration-tests/reproduces-transactions/BUILD.bazel
@@ -51,7 +51,6 @@ da_scala_test(
         "//libs-scala/fs-utils",
         "//libs-scala/ledger-resources",
         "//libs-scala/ports",
-        "//libs-scala/resources",
         "//test-common:dar-files-default-lib",
         "@maven//:io_netty_netty_handler",
         "@maven//:org_scalatest_scalatest_compatible",
