+++ daml-script/test/BUILD.bazel
@@ -156,7 +156,6 @@ da_scala_library(
         "//ledger/ledger-api-common",
         "//libs-scala/ledger-resources",
         "//libs-scala/ports",
-        "//libs-scala/resources",
         "//libs-scala/scala-utils",
         "@maven//:org_scalatest_scalatest_compatible",
     ],
@@ -215,6 +214,7 @@ da_scala_test_suite(
         "//ledger/participant-integration-api:participant-integration-api-tests-lib",
         "//ledger/participant-state",
         "//libs-scala/contextualized-logging",
+        "//libs-scala/fs-utils",
         "//libs-scala/jwt",
         "//libs-scala/ledger-resources",
         "//libs-scala/logging-entries",
