+++ daml-script/test/src/test-utils/com/daml/lf/engine/script/test/AbstractScriptTest.scala
@@ -99,7 +99,7 @@ trait AbstractScriptTest extends CantonFixture with AkkaBeforeAndAfterAll {
     )
     Runner.connect(
       participantParams = params,
-      tlsConfig = tlsConfig,
+      tlsConfig = config.ltsClientConfig,
       maxInboundMessageSize = maxInboundMessageSize,
     )
   }
