+++ triggers/tests/BUILD.bazel
@@ -117,7 +117,6 @@ da_scala_library(
         "//ledger/sandbox-on-x:sandbox-on-x-test-lib",
         "//libs-scala/ledger-resources",
         "//libs-scala/ports",
-        "//libs-scala/resources",
         "//test-common",
         "//triggers/runner:trigger-runner-lib",
         "@maven//:org_scalatest_scalatest_compatible",
@@ -289,7 +288,6 @@ da_scala_library(
         "//libs-scala/ledger-resources",
         "//libs-scala/logging-entries",
         "//libs-scala/ports",
-        "//libs-scala/resources",
         "//libs-scala/scala-utils",
         "//observability/tracing",
         "//triggers/runner:trigger-runner-lib",
