+++ triggers/tests/src/test/scala/com/digitalasset/daml/lf/engine/trigger/test/Jwt.scala
@@ -7,11 +7,10 @@ import akka.stream.scaladsl.Flow
 import com.daml.ledger.api.domain
 import com.daml.ledger.api.v1.commands.CreateCommand
 import com.daml.ledger.api.v1.{value => LedgerApi}
-import com.daml.lf.data.Ref
 import com.daml.lf.data.Ref._
 import com.daml.lf.engine.trigger.Runner.TriggerContext
 import com.daml.lf.engine.trigger.TriggerMsg
-import com.daml.lf.integrationtest.CantonFixture.{adminUserId, freshUserId}
+import com.daml.lf.integrationtest.CantonFixture
 import org.scalatest._
 import org.scalatest.matchers.should.Matchers
 import org.scalatest.wordspec.AsyncWordSpec
@@ -34,14 +33,14 @@ class Jwt extends AsyncWordSpec with AbstractTriggerTestWithCanton with Matchers
       )
     "1 create" in {
       for {
-        adminClient <- defaultLedgerClient(getToken(adminUserId))
-        userId = Ref.UserId.assertFromString(freshUserId())
+        adminClient <- defaultLedgerClient(config.adminToken)
+        userId = CantonFixture.freshUserId()
         partyDetails <- adminClient.partyManagementClient.allocateParty(None, None)
         party = partyDetails.party
         user = domain.User(userId, None)
         rights = Seq(domain.UserRight.CanActAs(party))
         _ <- adminClient.userManagementClient.createUser(user, rights)
-        client <- defaultLedgerClient(getToken(userId))
+        client <- defaultLedgerClient(config.getToken(userId))
         runner = getRunner(client, QualifiedName.assertFromString("ACS:test"), party)
         (acs, offset) <- runner.queryACS()
         // Start the future here
